% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_KBDI.R
\name{calculate_KBDI}
\alias{calculate_KBDI}
\title{Calculate KBDI (Mediterranean Version) for Drought Monitoring (Ganatsas et al., 2011)}
\usage{
calculate_KBDI(
  data,
  prec_col = "prec",
  temp_col = "Tmax",
  FC = 200,
  save_to_csv = FALSE,
  file_name = "KBDI_output.csv"
)
}
\arguments{
\item{data}{A daily climate data frame containing the input data.
It must include columns for 'year', 'month', 'day', daily precipitation, and daily maximum temperature.
The default columns are "prec" for precipitation and "Tmax" for maximum temperature, but these can be specified using `prec_col` and `temp_col` arguments.}

\item{prec_col}{A string specifying the name of the precipitation column. Default is "prec".}

\item{temp_col}{A string specifying the name of the temperature column. Default is "Tmax".}

\item{FC}{A numeric value representing the field capacity (in mm). Default is 200.}

\item{save_to_csv}{A logical value (`TRUE` or `FALSE`). If `TRUE`, the function will save the results to a CSV file. Default is `FALSE`.}

\item{file_name}{A string specifying the name of the CSV file to save the results if `save_to_csv` is `TRUE`. Default is "KBDI_output.csv".}
}
\value{
A data frame containing the original `year`, `month`, and `day` columns from the input data, along with the calculated `DF` (drying factor) and `KBDI` values for each day.
If `save_to_csv` is `TRUE`, the results will also be saved as a CSV file.
}
\description{
This function calculates the KBDI (Keetch-Byram Drought Index) for Mediterranean conditions, based on
daily precipitation (`prec`) and maximum temperature (`Tmax`) data. The KBDI is calculated using a time series
of daily data, and an optional argument allows saving the output to a CSV file.
It is highly recommended that users verify the input climate dataset with the provided checking functions `check_years_days_and_start` and `check_missing_values` before running the KBDI calculation.
This ensures that all days within each year are present with no missing values (NA), the years are in chronological order, and each year begins on January 1st.
}
\examples{
# Calculate KBDI for a dataset 'tab_clim' with columns 'prec' and 'Tmax'
result_kbdi <- calculate_KBDI(tab_clim)

# Calculate KBDI and save the results as a CSV file
result_kbdi_with_csv_saved <- calculate_KBDI(tab_clim, save_to_csv = TRUE, file_name = "KBDI_output.csv")

}
