% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns_names.R
\name{check_columns_names}
\alias{check_columns_names}
\title{Check and Print Column Names in the Daily Climate Dataset}
\usage{
check_columns_names(
  data,
  required_cols = c("year", "month", "day", "prec", "Tmax")
)
}
\arguments{
\item{data}{A data frame containing the daily climate dataset to be checked. The dataset should
include specific columns like 'year', 'month', 'day', 'prec', 'Tmax', but additional columns
are also allowed.}

\item{required_cols}{A vector of character strings representing the names of the
required columns. Default is `c('year', 'month', 'day', 'prec', 'Tmax')`.}
}
\value{
This function does not return a value but prints messages to the console.
If all required columns are present, it prints "Dataset has the correct columns"
If any required columns are missing, it prints "Dataset is missing some required columns"
It also prints any extra columns found in the dataset.
}
\description{
This function checks if all the required columns (e.g., 'year', 'month',
'day', 'prec', 'Tmax') are present in the input daily climate dataset. It also identifies any extra
columns that may be in the dataset. The function prints warnings if any required
columns are missing and prints messages for extra columns found. It returns a
message indicating if the dataset has the correct columns or is missing any required columns.
}
\details{
This function checks if the provided input daily climate dataset contains the required columns and prints
appropriate messages. It also reports any extra columns present in the input dataset.
}
\examples{
# Assuming 'tab_clim' is the daily input climate dataset
check_columns_names(tab_clim)

}
