% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_values.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check Missing Values in each column of the daily climate dataset}
\usage{
check_missing_values(data)
}
\arguments{
\item{data}{A daily climate data frame that will be checked for missing values (NA).}
}
\value{
This function does not return a value. It prints either a summary of missing values
per variable if missing values are detected, or a message indicating that no missing values are found.
}
\description{
This function checks whether any missing values (NA) are present in the input climate dataset.
 Before doing so, users should ensures that all years are in chronological order, contain either 365 or 366 (leap year) days, and start on January 1st, using the function `check_years_days_and_start`
 If missing values are found, it prints a summary of the number of missing values per variable.
 If no missing values are detected, it prints a confirmation message indicating that no missing values were found.
}
\examples{
# Check missing values in 'tab_clim' dataset
check_missing_values(tab_clim)


}
