% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_years_days_and_start.R
\name{check_years_days_and_start}
\alias{check_years_days_and_start}
\title{Check Yearly Day Counts and Starting Date}
\usage{
check_years_days_and_start(data)
}
\arguments{
\item{data}{A daily climate data frame containing the columns `year`, `month`, and `day`.
The function assumes that these columns represent the year, month, and day for each observation in the dataset.}
}
\value{
The function does not return a value, but prints messages to the console.
If all years have 365 or 366 days and start on January 1st, it prints a confirmation message.
If not, it prints which years are incomplete or do not start on January 1st. Also, if years are not
in a sequential order the function will stop and print a message about the displaced years.
}
\description{
This function checks three conditions for a given input Daily climate dataset:
1. Whether all years have exactly 365 or 366 days.
2. Whether all years start on January 1st.
3. Whether the year are in chronological order.
It prints messages indicating whether the conditions are met and provides details
on any discrepancies, including which years have fewer than 365 days and do not start on January 1st
}
\examples{
# Assuming 'tab_clim' is a dataset with 'year', 'month', and 'day' columns
check_years_days_and_start(tab_clim)

# Example where some years might have fewer than 365 days or don't start on January 1st
check_years_days_and_start(custom_data)

}
