% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drought_features_extraction.R
\name{Drought_features_extraction}
\alias{Drought_features_extraction}
\title{Yearly Drought Features Extraction}
\usage{
Drought_features_extraction(
  kbdi_data,
  minDmax,
  drought_features_table,
  save_to_csv = FALSE,
  plot_features_extraction = TRUE,
  file_name = "Drought_Features_table.csv"
)
}
\arguments{
\item{kbdi_data}{A data frame containing daily 'KBDI' data, and 'year', 'month', and 'day' columns, which are the output of the `calculate_KBDI` function.
It is crucial to ensure that all years are fully present, in chronological order, and without any missing KBDI values for any given day.}

\item{minDmax}{A data frame containing minDmax_DOY (in day since beginning) and its corresponding KBDI value (min_KBDI), which is the output of the `identify_min_kbdi_between_two_drought_peaks` function}

\item{drought_features_table}{an empty table for storing extracted drought features, which is the output of the `initialize_drought_features_table` function
As previously described, this function will exclude the last year in the provided climate dataset as we cannot characterize it due to lack of sufficient KBDI data on its subsequent year.}

\item{save_to_csv}{A logical value (`TRUE` or `FALSE`). If `TRUE`, the function will save the results to a CSV file. Default is `FALSE`.}

\item{plot_features_extraction}{A logical value (`TRUE` or `FALSE`). If `TRUE`, the function will plots the results of drought features extraction. Default is `TRUE`.}

\item{file_name}{A string specifying the name of the CSV file to save the results if `save_to_csv` is `TRUE`. Default is "Drought_Features_table.csv".}
}
\value{
The function returns a data frame populated with the extracted drought features for each hydrological year.
}
\description{
This function extracts 19 annual drought features for each hydrological year, starting on January 1st
and ending either when soil moisture is replenished to field capacity (as indicated by a minimum
KBDI near 0 between two peak drought days) or, if full replenishment is not reached, on the corresponding minDmax_DOY. Drought features extraction are based on three different soil desiccation thresholds,
where soil losses 25%, 50%, and 75% of its total available water content. The extracted features for
each year are saved in the pre-initialized table created by the `initialize_drought_features_table`function.
Detailed descriptions of each drought feature can be found in the foundational work by Elias et al., 2024.
Due to the incomplete return of soil moisture to field capacity at the end of some dry seasons under semi-arid and arid climatic conditions, Multi-Year Drought (MYD) events
can extend over consecutive hydrological years, possibly affecting various soil desiccation thresholds.
As a result, the output of this function may still contain missing values (NA) for some drought features that could not be calculated within those hydrological years.
Users should subsequently apply the MYD_characterization function to re-calculate and complete these missing features in the context of prolonged MYD events.
}
\examples{
# calculate KBDI
result_kbdi <- calculate_KBDI(tab_clim)

# calculate Peak.KBDI.DOY and its corresponding value Peak.KBDI over each hydrological year
peak_drought_info <- identify_peak_drought_days(result_kbdi)

# Identify Minimum KBDI day and corresponding value Between Two Consecutive Peak Drought Days (Peak.KBDI.DOY)
min_kbdi_between_drought_peaks <- identify_min_kbdi_between_two_drought_peaks(result_kbdi, peak_drought_info)

# Initialize an empty table to store the extracted drought features
drought_features_table  <- initialize_drought_features_table(result_kbdi)


# Drought features extraction
DF <- Drought_features_extraction (kbdi_data = result_kbdi,
                                   minDmax = min_kbdi_between_drought_peaks,
                                   drought_features_table = drought_features_table,
                                   save_to_csv = F,
                                   plot_features_extraction = T)

 head(DF)
}
