% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_min_kbdi_between_two_drought_peaks.R
\name{identify_min_kbdi_between_two_drought_peaks}
\alias{identify_min_kbdi_between_two_drought_peaks}
\title{Identify Minimum KBDI Value and Corresponding Day Since Beginning Between Two Consecutive Peak Drought Days (Peak.KBDI.DOY)}
\usage{
identify_min_kbdi_between_two_drought_peaks(kbdi_data, peak_drought_data)
}
\arguments{
\item{kbdi_data}{A data frame containing the daily calculated KBDI values in a column named "KBDI", in addition of 'year', 'month', and 'day' columns, which are the output of the `calculate_KBDI` function.
Users should ensure that all target years are fully represented, with each day included, and that all KBDI values contain no missing (NA) values.}

\item{peak_drought_data}{A data frame containing yearly peak KBDI days (DOY) and peak KBDI values,
which is the output from the `identify_peak_drought_days` function. It should include three columns:
"year", "Peak.KBDI.DOY", and "Peak.KBDI".}
}
\value{
A data frame with columns:
\item{year}{The year corresponding to each calculated minimum KBDI.}
\item{minDmax_DOY}{The day of the minimum KBDI between two consecutive peak days,
in days since the beginning of the KBDI time series.}
\item{min_KBDI}{The minimum KBDI value between two consecutive peak KBDI days.}
}
\description{
This function retrieve the minimum KBDI value and its corresponding day since the beginning of the KBDI time series,
particularly, between two consecutive peak KBDI days (Peak.KBDI.DOY).The retrieved 'minDmax_DOY' value will subsequently be used in drought features extraction,
serving to mark the end of consecutive hydrological years, which may occur after December 31st.
It's corresponding value 'min_KBDI' could be zero when the soil is fully replenished to its field capacity at the end of the dry season.
However, in the case of a MYD drought, the soil may not reach full replenishment, and the minimum 'min_KBDI' value could be greater than zero.
The function uses the peak drought days identified from the `identify_peak_drought_days` function to define the range for finding
the minimum KBDI value.
This function will return a table with the number of rows (equal to the number of years) reduced by one, as it cannot calculate the minDmax_DOY for the final year.
}
\examples{
# Example usage:
# Assuming `result_kbdi` is the output of the `calculate_KBDI` function, and `peak_drought_info` is the output of `identify_peak_drought_days` function
min_kbdi_between_drought_peaks <- identify_min_kbdi_between_two_drought_peaks(result_kbdi, peak_drought_info)


}
