% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_drought_features_table.R
\name{initialize_drought_features_table}
\alias{initialize_drought_features_table}
\title{Initialize Empty Drought Features Table}
\usage{
initialize_drought_features_table(kbdi_data)
}
\arguments{
\item{kbdi_data}{A data frame containing daily KBDI data, including a 'year' column, which is the output of the `calculate_KBDI` function}
}
\value{
A data frame with drought feature columns initialized for each hydrological year that will be characterized.
}
\description{
Creates an empty table for storing extracted drought features from daily calculated KBDI data for the considered years.
Users should ensure that the input data frame for this function contains all years fully represented and arranged in chronological order.
Additionally, this function excludes the last year in the provided data frame;
for instance, if daily KBDI values span from 1960 to 2020, the year 2020 will be excluded.
This exclusion occurs because drought features for 2020 cannot be extracted reliably without data from 2021, as the subsequent year's information is required for accurate calculation.
}
\examples{
# Initialize an empty table for drought feature extraction
drought_features_table  <- initialize_drought_features_table(result_kbdi)
head(drought_features_table)
}
